assignment_scrapping/requirements.txt
aiocontextvars==0.2.2
boto3==1.16.30
botocore==1.19.30
contextvars==2.4
environs==9.3.0
et-xmlfile==1.0.1
immutables==0.14
jdcal==1.4.1
jmespath==0.10.0
loguru==0.5.3
marshmallow==3.10.0
numpy==1.19.5
openpyxl==3.0.5
pandas==1.1.5
peewee==3.14.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
s3transfer==0.3.3
six==1.15.0
urllib3==1.26.2
xlrd==1.2.0
